applied custom Hooks and make it reusable
